+++ src/main/java/com/taskrail/controller/HomeController.java
@@ -2,8 +2,10 @@ package com.taskrail.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
+@RequestMapping("/view")
 public class HomeController {
 
     // 커버 페이지
